updated hooks, added another component updated app
