+++ src/Components/Nav.js
@@ -1,8 +1,5 @@
 import React, {useState} from 'react';
-import MyYards from './MyYards';
-import HostYardListing from './HostYard';
-import UserLogin from './UserLogin.js';
-import NewHost from './NewHost';
+
 // import Container from 'react-bootstrap/Container';
 // import logo from "../assets/yardMeLogo.png"
 import Nav from 'react-bootstrap/Nav';
@@ -16,36 +13,36 @@ import NavDropdown from 'react-bootstrap/NavDropdown';
 
 
 
-  function YardNav() {
-    const [page, setPage] = useState("userLogin")
-        function switchPage () {
-            if (page === "MyYards") {
-                return (
-                <MyYards/>
+  function YardNav({setPage, setIsLoggedIn}) {
+//     const [page, setPage] = useState("userLogin")
+//         function switchPage () {
+//             if (page === "MyYards") {
+//                 return (
+//                 <MyYards/>
                
-             )
-        } else if (page === "HostYard") {
-                return (
-            <div>
-                 <HostYardListing/>
-            </div>
-        )
-      }
-      else if (page === "UserLogin") {
-        return (
-    <div>
-         <UserLogin/>
-    </div>
-)
-  }
-  else if (page === "NewHost") {
-    return (
-<div>
-     <NewHost/>
-</div>
-)
-}
-};
+//              )
+//         } else if (page === "HostYard") {
+//                 return (
+//             <div>
+//                  <HostYardListing/>
+//             </div>
+//         )
+//       }
+//       else if (page === "UserLogin") {
+//         return (
+//     <div>
+//          <UserLogin/>
+//     </div>
+// )
+//   }
+//   else if (page === "NewHost") {
+//     return (
+// <div>
+//      <NewHost/>
+// </div>
+// )
+// }
+// };
   
     return (
       <div> 
@@ -59,11 +56,11 @@ import NavDropdown from 'react-bootstrap/NavDropdown';
             {/* <NavDropdown.Item href="#action/3.1">HostYard</NavDropdown.Item>
             <NavDropdown.Item  href="#action/3.2">HostNew</NavDropdown.Item> */}
              {/* <NavDropdown.Item href="#action/3.3">LogIn</NavDropdown.Item> */}
-            <NavDropdown.Item><span onClick = {() => setPage('UserLogin')}>LogOut</span></NavDropdown.Item>
+            <NavDropdown.Item><span onClick = {() => setIsLoggedIn(false)}>LogOut</span></NavDropdown.Item>
           </NavDropdown>
         </Nav>
       </Navbar.Collapse>
-      {switchPage()}
+      {/* {switchPage()} */}
       </div>
     )
   };
